refactor(app): more device hooks from pages
