+++ app/src/organisms/ApplyHistoricOffsets/LabwareOffsetTable.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { useTranslation } from 'react-i18next'
 import { SPACING, TYPOGRAPHY, COLORS } from '@opentrons/components'
 import { OffsetVector } from '/app/molecules/OffsetVector'
-import { formatTimestamp } from '../Devices/utils'
+import { formatTimestamp } from '/app/transformations/runs'
 import { getDisplayLocation } from '../LabwarePositionCheck/utils/getDisplayLocation'
 import type { LabwareDefinition2 } from '@opentrons/shared-data'
 import type { OffsetCandidate } from './hooks/useOffsetCandidatesForAnalysis'
