+++ app/src/organisms/Breadcrumbs/__tests__/Breadcrumbs.test.tsx
@@ -6,7 +6,7 @@ import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { fireEvent, screen } from '@testing-library/react'
 
 import { i18n } from '/app/i18n'
-import { useRunCreatedAtTimestamp } from '/app/organisms/Devices/hooks'
+import { useRunCreatedAtTimestamp } from '/app/resources/runs'
 import { useRobot } from '/app/redux-resources/robots'
 import { getProtocolDisplayName } from '/app/organisms/ProtocolsLanding/utils'
 import { getIsOnDevice } from '/app/redux/config'
@@ -18,7 +18,7 @@ import { Breadcrumbs } from '..'
 
 import type { State } from '/app/redux/types'
 
-vi.mock('/app/organisms/Devices/hooks')
+vi.mock('/app/resources/runs')
 vi.mock('/app/redux-resources/robots')
 vi.mock('/app/organisms/ProtocolsLanding/utils')
 vi.mock('/app/redux/config')
