+++ app/src/organisms/Breadcrumbs/index.tsx
@@ -18,7 +18,7 @@ import {
 } from '@opentrons/components'
 import { ApiHostProvider } from '@opentrons/react-api-client'
 
-import { useRunCreatedAtTimestamp } from '/app/organisms/Devices/hooks'
+import { useRunCreatedAtTimestamp } from '/app/resources/runs'
 import { getProtocolDisplayName } from '/app/organisms/ProtocolsLanding/utils'
 import { getIsOnDevice } from '/app/redux/config'
 import { OPENTRONS_USB } from '/app/redux/discovery'
