+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/RunHeaderSectionUpper.tsx
@@ -13,8 +13,7 @@ import { LabeledValue } from './LabeledValue'
 import { DisplayRunStatus } from '../DisplayRunStatus'
 import { RunTimer } from '/app/molecules/RunTimer'
 import { ActionButton } from './ActionButton'
-import { useRunCreatedAtTimestamp } from '../../../hooks'
-import { useRunTimestamps } from '/app/resources/runs'
+import { useRunTimestamps, useRunCreatedAtTimestamp } from '/app/resources/runs'
 
 import type { RunHeaderContentProps } from '.'
