+++ app/src/organisms/Devices/ProtocolRun/__tests__/SetupCalibrationItem.test.tsx
@@ -6,11 +6,11 @@ import { describe, it, beforeEach, vi, afterEach } from 'vitest'
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { useRunHasStarted } from '/app/resources/runs'
-import { formatTimestamp } from '../../utils'
+import { formatTimestamp } from '/app/transformations/runs'
 import { SetupCalibrationItem } from '../SetupCalibrationItem'
 
 vi.mock('/app/resources/runs')
-vi.mock('../../utils')
+vi.mock('/app/transformations/runs')
 
 const RUN_ID = '1'
