+++ app/src/organisms/Devices/__tests__/utils.test.tsx
@@ -1,7 +1,6 @@
 import { describe, it, expect } from 'vitest'
 import '@testing-library/jest-dom/vitest'
 import {
-  formatTimestamp,
   getIs96ChannelPipetteAttached,
   getOffsetCalibrationForMount,
 } from '../utils'
@@ -14,28 +13,6 @@ import type {
   FetchPipettesResponseBody,
   FetchPipettesResponsePipette,
 } from '/app/redux/pipettes/types'
-
-describe('formatTimestamp', () => {
-  it('should format an ISO 8601 date string', () => {
-    const date = '2021-03-07T18:44:49.366581+00:00'
-
-    expect(formatTimestamp(date)).toMatch(
-      /^(\d{2})\/(\d{2})\/(\d{4}) (\d{2}):(\d{2}):(\d{2})$/
-    )
-  })
-
-  it('should pass through a non-ISO 8601 date string', () => {
-    const date = '2/22/2022 1:00'
-
-    expect(formatTimestamp(date)).toEqual(date)
-  })
-
-  it('should pass through a non-date string', () => {
-    const noDate = 'A Protocol For Otie'
-
-    expect(formatTimestamp(noDate)).toEqual(noDate)
-  })
-})
 
 describe('getIs96ChannelPipetteAttached hook', () => {
   it('returns false when there is no pipette attached on the left mount', () => {
