+++ app/src/organisms/Devices/hooks/index.ts
@@ -11,7 +11,6 @@ export * from './useModuleRenderInfoForProtocolById'
 export * from './useModuleCalibrationStatus'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
-export * from './useRunCreatedAtTimestamp'
 export * from './useTipLengthCalibrations'
 export * from './useUnmatchedModulesForProtocol'
 export * from './useProtocolAnalysisErrors'
