+++ app/src/resources/runs/__tests__/useRunCreatedAtTimestamp.test.tsx
@@ -3,15 +3,15 @@ import { vi, it, expect, describe, beforeEach } from 'vitest'
 import { when } from 'vitest-when'
 
 import { mockIdleUnstartedRun } from '/app/resources/runs/__fixtures__'
-import { formatTimestamp } from '../../utils'
+import { formatTimestamp } from '/app/transformations/runs'
 import { useRunCreatedAtTimestamp } from '../useRunCreatedAtTimestamp'
-import { useNotifyRunQuery } from '/app/resources/runs'
+import { useNotifyRunQuery } from '../useNotifyRunQuery'
 
 import type { UseQueryResult } from 'react-query'
 import type { Run } from '@opentrons/api-client'
 
-vi.mock('/app/resources/runs')
-vi.mock('../../utils')
+vi.mock('../useNotifyRunQuery')
+vi.mock('/app/transformations/runs')
 
 const MOCK_RUN_ID = '1'
