+++ app/src/resources/runs/index.ts
@@ -21,3 +21,4 @@ export * from './useMostRecentCompletedAnalysis'
 export * from './useRunPipetteInfoByMount'
 export * from './useRunCalibrationStatus'
 export * from './useRunHasStarted'
+export * from './useRunCreatedAtTimestamp'
