+++ app/src/transformations/runs/formatTimestamp.ts
@@ -1,0 +1,12 @@
+import { format, parseISO } from 'date-fns'
+/**
+ * formats a string if it is in ISO 8601 date format
+ * @param {string} timestamp ISO date string
+ * @returns {string} formatted date string
+ */
+export function formatTimestamp(timestamp: string): string {
+    // eslint-disable-next-line eqeqeq
+    return (parseISO(timestamp) as Date | string) != 'Invalid Date'
+        ? format(parseISO(timestamp), 'MM/dd/yyyy HH:mm:ss')
+        : timestamp
+}
